ntt-cv-builder/requirements.txt
# Web Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
websockets==13.1
aiofiles==24.1.0

# OpenAI
openai==1.55.3
openai-agents==0.0.9

# LangChain + RAG
langchain==0.3.7
langchain-openai==0.2.9
langchain-chroma==0.1.4
langchain-community==0.3.7
chromadb==0.5.18

# Data Validation
pydantic==2.10.1
pydantic-settings==2.6.1

# File Parsing
pymupdf==1.24.13
python-docx==1.1.2
python-pptx==1.0.2

# Output Generation
weasyprint==62.3
jinja2==3.1.4
docxtpl==0.17.0

# Session / Caching
redis==5.2.0
cachetools==5.5.0

# Utilities
python-dotenv==1.0.1
httpx==0.27.2
tenacity==9.0.0
structlog==24.4.0
rich==13.9.4

# Dev / Testing
pytest==8.3.3
pytest-asyncio==0.24.0
